Fixed an issue in the make_guess function in main.py where the an error would be shown as I try to subscript an object rather than the dictionary within it.
Added a return to the comparison_result method of the PlayerDatabase class which would return a boolean representing whether the guess is correct or not.

Replaced the print placeholders for a successful guess made with code that runs the comparison_result method of the PlayerDatabase class.

Made it so that the make_guess function is run when the user types "Guess" for the initial prompt.