Added a tries input to the make_guess function in the main.py file which would allow for the input of the current number of attempts. Added code functionality to increment the number of tries by 1 per guess and added a return of this value.
Moved the csv file reading, random player generation and tries variable creation to outside of the prog_start loop as I do not want this to be done multiple times.

When the user types Guess for the initial prompt, the make_guess function's output will update the tries variable alongside adding to the correct_guess variable and a print is displayed when the correct_guess variable is True, representing a correct guess.