Moved the make_guess function to the utilities.py file
Added the check_players function to the main.py file which will take an input of 1 letter and list characters whose names/nickname start with that letter. Used to display characters to the user that can be entered.

Added code to the main prompt to run this function when the user types "check".