Added a few files to gitignore that are not necessary for the program.
Added the time import to main.py just in case I wished to use it.

Added the show_guesses function to main.py which takes the styled guess results and prints them to the console.

Changed the console width to 120.

Added a section to main.py that stores styled guess results and appended the styled header info to this section.

Added a section to main,py that appends styled guess results created by the make_guess function and runs the show_guesses function.