Added the show_checks function to the main.py file which will print the characters not guessed beginning with the letter specified via the check option.
Added code to run this function if there are results from a check that have appeared (to give users a reference to the characters they checked for).

Changed the check part of the main function to reflect updates I've made to the check_players function.

Changed the check_players function in utilities.py to return a list of styled strings that will be printed to console via the show_checks function rather than within the check_players function.